+++ actix-files/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- `PathBufWrap` & `UriSegmentError` made public.
 - Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.6.6