+++ actix-files/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.6.7
+
 - Add `{Files, NamedFile}::read_mode_threshold()` methods to allow faster synchronous reads of small files.
 - Minimum supported Rust version (MSRV) is now 1.75.